Refactor arg parser into separate files
